refactor: add pages directory and split layout
